Add page object to capture common tasks in dashboard detail page
